+++ modules/kubernetes/azure-metrics/main.tf
@@ -9,6 +9,10 @@ terraform {
   required_version = ">= 1.3.0"
 
   required_providers {
+    azurerm = {
+      version = "3.99.0"
+      source  = "hashicorp/azurerm"
+    }
     git = {
       source  = "xenitab/git"
       version = "0.0.3"
@@ -19,16 +23,16 @@ terraform {
 resource "git_repository_file" "kustomization" {
   path = "clusters/${var.cluster_id}/azure-metrics.yaml"
   content = templatefile("${path.module}/templates/kustomization.yaml.tpl", {
-    cluster_id = var.cluster_id
+    cluster_id = var.cluster_id,
   })
 }
 
 resource "git_repository_file" "azure_metrics" {
   path = "platform/${var.cluster_id}/azure-metrics/azure-metrics.yaml"
   content = templatefile("${path.module}/templates/azure-metrics.yaml.tpl", {
-    client_id               = var.client_id
-    subscription_id         = var.subscription_id
-    podmonitor_kubernetes   = var.podmonitor_kubernetes
-    podmonitor_loadbalancer = var.podmonitor_loadbalancer
+    client_id               = azurerm_user_assigned_identity.azure_metrics.client_id,
+    subscription_id         = var.subscription_id,
+    podmonitor_kubernetes   = var.podmonitor_kubernetes,
+    podmonitor_loadbalancer = var.podmonitor_loadbalancer,
   })
 }