+++ modules/kubernetes/cert-manager/main.tf
@@ -8,6 +8,10 @@ terraform {
   required_version = ">= 1.3.0"
 
   required_providers {
+    azurerm = {
+      version = "3.99.0"
+      source  = "hashicorp/azurerm"
+    }
     git = {
       source  = "xenitab/git"
       version = "0.0.3"
@@ -15,23 +19,25 @@ terraform {
   }
 }
 
-locals {
-  azure_hosted_zone_names = "[${join(",", var.azure_config.hosted_zone_names)}]"
-}
+#locals {
+#  azure_hosted_zone_names = "[${join(",", var.dns_zones)}]"
+#}
 
 resource "git_repository_file" "kustomization" {
   path = "clusters/${var.cluster_id}/cert-manager.yaml"
   content = templatefile("${path.module}/templates/kustomization.yaml.tpl", {
-    cluster_id = var.cluster_id
+    cluster_id = var.cluster_id,
   })
 }
 
 resource "git_repository_file" "cert_manager" {
   path = "platform/${var.cluster_id}/cert-manager/cert-manager.yaml"
   content = templatefile("${path.module}/templates/cert-manager.yaml.tpl", {
-    notification_email = var.notification_email,
-    acme_server        = var.acme_server,
-    azure_config       = var.azure_config,
-    dns_zones          = local.azure_hosted_zone_names
+    acme_server         = var.acme_server,
+    client_id           = azurerm_user_assigned_identity.cert_manager.client_id,
+    dns_zones           = var.dns_zones,
+    notification_email  = var.notification_email,
+    resource_group_name = var.global_resource_group_name,
+    subscription_id     = var.subscription_id,
   })
 }