+++ modules/kubernetes/prometheus/variables.tf
@@ -1,21 +1,21 @@
+variable "aks_name" {
+  description = "The AKS cluster short name, e.g. 'aks'."
+  type        = string
+}
+
 variable "azure_config" {
   description = "Azure specific configuration"
   type = object({
     azure_key_vault_name = string
-    identity = object({
-      client_id   = string
-      resource_id = string
-      tenant_id   = string
-    })
   })
   default = {
     azure_key_vault_name = ""
-    identity = {
-      client_id   = ""
-      resource_id = ""
-      tenant_id   = ""
-    }
   }
+}
+
+variable "cluster_id" {
+  description = "Unique identifier of the cluster across regions and instances."
+  type        = string
 }
 
 variable "cluster_name" {
@@ -28,19 +28,24 @@ variable "environment" {
   type        = string
 }
 
+variable "location_short" {
+  description = "The Azure region short name."
+  type        = string
+}
+
+variable "namespace_selector" {
+  description = "Kind labels to look for in namespaces"
+  type        = list(string)
+  default     = ["platform"]
+}
+
+variable "oidc_issuer_url" {
+  description = "Kubernetes OIDC issuer URL for workload identity."
+  type        = string
+}
+
 variable "region" {
   description = "The region in which the prometheus instance is deployed"
-  type        = string
-}
-
-variable "tenant_id" {
-  description = "The tenant id label to apply to all metrics in remote write"
-  type        = string
-  default     = ""
-}
-
-variable "cluster_id" {
-  description = "Unique identifier of the cluster across regions and instances."
   type        = string
 }
 
@@ -55,6 +60,17 @@ variable "remote_write_url" {
   type        = string
 }
 
+variable "resource_group_name" {
+  description = "The Azure resource group name"
+  type        = string
+}
+
+variable "resource_selector" {
+  description = "Monitoring type labels to look for in Prometheus resources"
+  type        = list(string)
+  default     = ["platform"]
+}
+
 variable "volume_claim_storage_class_name" {
   description = "StorageClass name that your pvc will use"
   type        = string
@@ -65,18 +81,6 @@ variable "volume_claim_size" {
   description = "Size of prometheus disk"
   type        = string
   default     = "10Gi"
-}
-
-variable "resource_selector" {
-  description = "Monitoring type labels to look for in Prometheus resources"
-  type        = list(string)
-  default     = ["platform"]
-}
-
-variable "namespace_selector" {
-  description = "Kind labels to look for in namespaces"
-  type        = list(string)
-  default     = ["platform"]
 }
 
 # Opt in to additional monitors