+++ validation/kubernetes/trivy/main.tf
@@ -3,8 +3,14 @@ terraform {}
 module "trivy" {
   source = "../../../modules/kubernetes/trivy"
 
-  client_id                       = "id"
-  cluster_id                      = "id"
-  resource_id                     = "id"
+  aks_managed_identity            = "id"
+  aks_name                        = "aks"
+  cluster_id                      = "foo"
+  environment                     = "dev"
+  location                        = "location"
+  location_short                  = "we"
+  oidc_issuer_url                 = "url"
+  resource_group_name             = "rg-name"
+  unique_suffix                   = "1234"
   volume_claim_storage_class_name = "name"
 }